+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.0.3
 
+### Added
+
+* Added build.ps1 and TMDeepSecurity.build.ps1 for build process
+
 ### Changed
 
-* 
+* Updated Help files